+++ src/components/tree/tree.js
@@ -30,6 +30,7 @@ const COMPONENT_NAME = 'tree';
 * @param {null|function} [settings.onBeforeSelect] If defined as a function, fires that function as a callback before the selection on a node occurs.
 * @param {null|function} [settings.onExpand] If defined as a function, fires that function as a node is expanded.
 * @param {null|function} [settings.onCollapse] If defined as a function, fires that function as a node is collapsed.
+* @param {boolean} [settings.toggleByIcon = true] if `true` && selectable set to `single`, this will allows to toggle parent nodes only by click on icon.
 */
 
 const TREE_DEFAULTS = {
@@ -44,7 +45,8 @@ const TREE_DEFAULTS = {
   sortable: false, // Allow nodes to be sortable
   onBeforeSelect: null,
   onExpand: null,
-  onCollapse: null
+  onCollapse: null,
+  toggleByIcon: false
 };
 
 function Tree(element, settings) {
@@ -476,7 +478,7 @@ Tree.prototype = {
    */
   selectNodeFinish(node, focus, e) {
     // Don't do selection for toggle type only
-    if (this.isMultiselect && e) {
+    if ((this.isMultiselect || this.settings.toggleByIcon) && e) {
       if (e.type === 'click' || e.type === 'touch') {
         if (DOM.hasClass(e.target, 'icon') &&
           node[0].parentNode.classList.contains('folder')) {
@@ -882,7 +884,7 @@ Tree.prototype = {
       const parent = this.parentNode;
       utils.clearSelection(); // Deselect all selected text.
       if (!target[0].classList.contains('is-disabled') && !target[0].classList.contains('is-loading')) {
-        if (self.isMultiselect) {
+        if (self.isMultiselect || self.settings.toggleByIcon) {
           if (DOM.hasClass(e.target, 'icon') && parent.classList.contains('folder')) {
             self.toggleNode(target, e);
           } else if (parent.classList.contains('is-selected') || parent.classList.contains('is-partial')) {