+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `[Datagrid]` Added a setting disableRowDeselection that if enabled does not allow selected rows to be toggled to deselected. ([#3791](https://github.com/infor-design/enterprise/issues/3791))
 - `[Demoapp]` Added the ability to set runtime flags for persisting settings that were previously only possible to set via URL query parameters. ([n/a])
+- `[Tree]` Added support to toggle by clicked on icon only. ([#3730](https://github.com/infor-design/enterprise/issues/3730))
 
 ### v4.30.0 Fixes